+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## master
 
 - better behaviour for image interpretations like MATRIX
+- window menu has next-image / prev-image in the top line [euzada]
 
 # 3.0.4 1/3/24